add admin right click
