+++ BaseModel/src/com/wpi/cs509/teamA/ui/ImageMouseListener.java
@@ -21,7 +21,7 @@ public class ImageMouseListener implements MouseListener {
         stateContext.execute(e);
     }
 
-    
+
     @Override
     public void mousePressed(MouseEvent e) {
     }
