+++ BaseModel/src/com/wpi/cs509/teamA/ui/MouseActionAdminUser.java
@@ -1,5 +1,7 @@
 package com.wpi.cs509.teamA.ui;
 
+import javafx.scene.input.MouseButton;
+
 import java.awt.event.MouseEvent;
 
 /**
@@ -21,6 +23,15 @@ public class MouseActionAdminUser extends MouseActionState {
 
     @Override
     public boolean execute(MouseEvent e) {
+        xPos = e.getX();
+        yPos = e.getY();
+        if (e.getButton() == MouseEvent.BUTTON3 )
+        {
+            NodeManageMenu nodeManageMenu = new NodeManageMenu(stateContext.getImageComponent(), xPos, yPos, stateContext);
+            nodeManageMenu.show(e.getComponent(), xPos, yPos);
+            return false;
+        }
+
         return false;
     }
 }
